Change index integer to not be kind int64, so that set_det_diagh can be called.

+++ src/semi_stoch_procs.F90
@@ -543,7 +543,7 @@ contains
 
     subroutine fill_in_diag_helements()
 
-        integer(int64) :: i
+        integer :: i
         integer :: nI(nel)
         real(dp) :: tmpH
